requirements.txt
google.generativeai
openpyxl==3.1.5